ISSUE-173: Refactor even more classes
