+++ app/build.gradle
@@ -7,8 +7,8 @@ android {
         applicationId "ar.com.tzulberti.testing"
         minSdkVersion 16
         targetSdkVersion 25
-        versionCode 5
-        versionName "0.4.0"
+        versionCode 6
+        versionName "0.6.0"
         testInstrumentationRunner "android.support.test.runner.AndroidJUnitRunner"
     }
     buildTypes {
@@ -36,6 +36,6 @@ dependencies {
     compile 'com.jsibbold:zoomage:1.1.0'
     compile 'com.edmodo:rangebar:1.0.0'
     compile 'io.sentry:sentry-android:1.0.0'
-    compile 'com.amitshekhar.android:debug-db:1.0.0'
+    testCompile 'com.amitshekhar.android:debug-db:1.0.0'
     testCompile 'junit:junit:4.12'
 }
