--- app/src/main/res/layout/activity_main.xml
@@ -1,53 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<android.support.v4.widget.DrawerLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/drawer_layout"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    android:fitsSystemWindows="true"
-    tools:openDrawer="start">
-
-    <Button
-        android:id="@+id/main_activity_series"
-        android:text="@string/series"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:onClick="goToMainOption"
-        />
-
-    <Button
-        android:id="@+id/main_activity_today_retentions"
-        android:text="@string/leftMenuRetentions"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:onClick="goToMainOption"
-        />
-
-
-    <Button
-        android:id="@+id/main_activity_today_tournament"
-        android:text="@string/leftMenuTournament"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:onClick="goToMainOption"
-        />
-
-
-    <Button
-        android:id="@+id/main_activity_playoff"
-        android:text="@string/playoff"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:onClick="goToMainOption"
-        />
-
-    <Button
-        android:id="@+id/main_activity_bow"
-        android:text="@string/bow"
-        android:layout_height="wrap_content"
-        android:layout_width="match_parent"
-        android:onClick="goToMainOption"
-        />
-
-
-</android.support.v4.widget.DrawerLayout>
